Renamed and refactored script
